gamepack_backend/requirements.txt
# requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
dj-database-url
Pillow
gunicorn